rename(ReservationTime): Time -> ReservationTime 으로 수정
