+++ src/main/java/roomescape/reservationtime/repository/TimeRepository.java
@@ -1,10 +1,10 @@
-package roomescape.time.repository;
+package roomescape.reservationtime.repository;
 
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.repository.Repository;
-import roomescape.time.domain.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 
 public interface TimeRepository extends Repository<ReservationTime, Long> {
